update baskets and products model

+++ database/migrations/1551909503397_products_schema.js
@@ -7,7 +7,7 @@ class ProductsSchema extends Schema {
   up () {
     this.table('products', (table) => {
       // alter table
-      table.integer('cost').notNullable().defaultTo(0)
+      table.float('cost').notNullable().defaultTo(0)
     })
   }
